Add links to timestamp standards
Wikipedia has been linked to for ISO 8601 because www.iso.org does not
provide the standard free of charge [1] and the Wikipedia page does a good
job of describing it.

The index.html has not bee regenerated because doing so with the latest
version of bikeshed results in a lot of unrelated changes.

[1] http://www.iso.org/iso/home/standards/iso8601.htm